fix: Build error 문제 수정
