+++ feature/daily-manage/build.gradle.kts
@@ -6,11 +6,12 @@ plugins {
 }
 
 android {
-    namespace = "com.teampatch.daily_manage"
+    namespace = "com.teampatch.daily.manage"
 }
 
 dependencies {
 
+    implementation(project(":core:common"))
     implementation(project(":core:domain"))
     implementation(project(":core:designsystem"))
