+++ feature/daily-manage/src/main/java/com/teampatch/daily/manage/model/AnswerEvent.kt
@@ -1,4 +1,4 @@
-package com.teampatch.daily_manage.model
+package com.teampatch.daily.manage.model
 
 internal enum class AnswerEvent {
     EDIT
