+++ feature/daily-manage/src/main/java/com/teampatch/daily/manage/model/DailyManageUiState.kt
@@ -1,17 +1,12 @@
-package com.teampatch.daily_manage.model
+package com.teampatch.daily.manage.model
 
 import androidx.paging.PagingData
 import com.teampatch.core.domain.emptyUser
 import com.teampatch.core.domain.model.Daily
 import com.teampatch.core.domain.model.DailyComment
-import com.teampatch.core.domain.model.DailyManage
-import com.teampatch.core.domain.model.Question
-import com.teampatch.core.domain.model.QuestionComment
-import com.teampatch.core.domain.model.QuestionDetail
 import com.teampatch.core.domain.model.User
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.emptyFlow
-import java.time.LocalDateTime
 
 internal data class DailyManageUiState(
 //    val daily: Flow<PagingData<Daily>> = emptyFlow(),
